Add files via upload
GUESSING GAME:-
The guessing game is a simple interactive game where the player tries to guess a randomly generated number within a specified range. The game starts by generating a random number between a minimum and maximum value, such as 1 and 100. The player is then prompted to enter their guess.
The program compares the player's guess with the generated number and provides feedback to the player. If the guess is higher than the generated number, the program informs the player to guess lower. Conversely, if the guess is lower, the player is advised to guess higher. This feedback helps the player narrow down their subsequent guesses.

The game continues until the player guesses the correct number. Once the player makes the correct guess, the program congratulates them and displays the number of attempts it took to find the answer. Optionally, the program may offer the player to play again or exit the game.
Overall, the guessing game is an entertaining way to test a player's intuition and luck, as they attempt to determine the randomly generated number within a specified range using feedback from the program.